Restructured the app
